backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
google-genai
python-dateutil
python-multipart